Include Sigmoid Algo (#5)
* Update requirements

* Optional logging to file

* Add minimization to windmill test

* Add sigmoid algo

* Update and fix datapan

* Docstrings

---------

+++ utils/logging.py
@@ -1,8 +1,9 @@
 import logging
+import typing as t
 from pathlib import Path
 
 
-def setup_logger(path_save: Path):
+def setup_logger(path_save: t.Optional[Path] = None):
     """
     Setup logging both to console and the file.
 
@@ -18,11 +19,12 @@ def setup_logger(path_save: Path):
     handler_stream.setFormatter(formatter)
     logger.addHandler(handler_stream)
 
-    if path_save.exists():
-        path_save.unlink()
-    handler_file = logging.FileHandler(path_save, encoding="utf-8")
-    handler_file.setLevel(logging.INFO)
-    handler_file.setFormatter(formatter)
-    logger.addHandler(handler_file)
+    if path_save is not None:
+        if path_save.exists():
+            path_save.unlink()
+        handler_file = logging.FileHandler(path_save, encoding="utf-8")
+        handler_file.setLevel(logging.INFO)
+        handler_file.setFormatter(formatter)
+        logger.addHandler(handler_file)
